+++ python/hpctl/hpctl/__init__.py
@@ -1,0 +1,1 @@
+from hpctl.version import __version__
