+++ python/hpctl/hpctl/utils.py
@@ -73,100 +73,8 @@ class partialmethod(partial):
 
 
 @export
-def order_json(data):
-    """Sort json to a consistent order.
-
-    When you hash json that has the some content but is different orders you get
-    different fingerprints.
-
-    In:  hashlib.sha1(json.dumps({'a': 12, 'b':14}).encode('utf-8')).hexdigest()
-    Out: '647aa7508f72ece3f8b9df986a206d95fd9a2caf'
-    In:  hashlib.sha1(json.dumps({'b': 14, 'a':12}).encode('utf-8')).hexdigest()
-    Out: 'a22215982dc0e53617be08de7ba9f1a80d232b23'
-
-    This function sorts json by key so that hashes are consistent.
-
-    Note:
-        In our configs we only have lists where the order doesn't matter so we
-        can sort them for consistency. This would have to change if we add a
-        config field that needs order we will need to refactor this.
-
-    :param data: dict, The json data.
-
-    :returns:
-        collections.OrderedDict: The data in a more consistent order.
-    """
-    new = OrderedDict()
-    for key, value in sorted(data.items(), key=lambda x: x[0]):
-        if isinstance(value, dict):
-            value = order_json(value)
-        elif isinstance(value, list):
-            value = sorted(value)
-        new[key] = value
-    return new
-
-
-KEYS = {
-    ('conll_output',),
-    ('visdom',),
-    ('visdom_name',),
-    ('model', 'gpus'),
-    ('test_thresh',),
-    ('reporting',),
-    ('num_valid_to_show',),
-    ('train', 'verbose'),
-    ('train', 'model_base'),
-    ('train', 'model_zip'),
-    ('test_batchsz')
-}
-
-
-@export
-def remove_monitoring(config, keys=KEYS):
-    """Remove config items that don't effect the model.
-
-    When base most things off of the sha1 hash of the model configs but there
-    is a problem. Some things in the config file don't effect the model such
-    as the name of the `conll_output` file or if you are using `visdom`
-    reporting. This strips out these kind of things so that as long as the model
-    parameters match the sha1 will too.
-
-    :param config: dict, The json data.
-    :param keys: Set[Tuple[str]], The keys to remove.
-
-    :returns:
-        dict, The data with certain keys removed.
-    """
-    c = deepcopy(config)
-    for key in keys:
-        x = c
-        for k in key[:-1]:
-            x = x.get(k)
-            if x is None:
-                break
-        else:
-            _ = x.pop(key[-1], None)
-    return c
-
-
-@export
-def hash_config(config):
-    """Hash a json config with sha1.
-
-    :param config: dict, The config to hash.
-
-    :returns:
-        str, The sha1 hash.
-    """
-    stripped_config = remove_monitoring(config)
-    sorted_config = order_json(stripped_config)
-    json_bytes = json.dumps(sorted_config).encode('utf-8')
-    return hashlib.sha1(json_bytes).hexdigest()
-
-
-@export
 def create_logs(label, mead_logs, hpctl_logs):
-    """Inject the hpctl looging config into the mead logging.
+    """Inject the hpctl logging config into the mead logging.
 
     Note:
         This currently writes it a file that mead need for now.
